Use 64-bit recursion and proof lengths in verifier APIs.
Align the Python and C bindings with the native verifier interface by accepting full-size proof blob lengths and recursion depth instead of narrowing to int32, removing wrapper-level truncation guards.

+++ src/c_bindings/c_wrapper.cpp
@@ -44,10 +44,10 @@ extern "C" {
                 -discriminant,
                 x_s,
                 proof_blob,
-                checked_cast<int32_t>(proof_blob_size),
+                proof_blob_size,
                 num_iterations,
                 discriminant_size * 8,
-                checked_cast<int32_t>(recursion)
+                recursion
             );
         } catch (...) {
             return false;
